use "double" instead of "long double"
git-svn-id: https://pure-data.svn.sourceforge.net/svnroot/pure-data/trunk@2835 df7eb9d6-776f-4174-a42d-0ffd8d5d5b5d

+++ externals/iem/iemmatrix/src/iemmatrix.h
@@ -50,7 +50,7 @@
 # define fabsf fabs
 #endif
 
-typedef long double t_matrixfloat;
+typedef double t_matrixfloat;
 
 
 /* the main class...*/
